Complete audit remediation: implement 75 deferred findings
Second remediation pass over the audit backlog. Cumulative across all waves:
247 of 261 findings now resolved; build verified green.

Notable in this pass:
- Security: imported instance bundles now strip untrusted hooks and tool
  command overwrites (prevents arbitrary-command execution on import);
  fuse-overlayfs commands POSIX-escaped; NexusMods dialog HTML-escapes all
  network strings; deploy hook fields documented as shell-interpreted.
- Robustness: deployer-index bounds guards; RAII profile restore in
  exportProfile; atomic deploy via same-dir temp + rename; cross-device
  setStagingDir fallback with rollback; error_code-based deleteAllData;
  group-erase lockstep helper; OpenMW undeploy backup restore; persistent
  log stream; LSPAK extraction size caps; release.yml linuxdeploy pinning +
  checksum verification.

Remaining 14 (1 High: runCommand->QProcess; plus KDF, CI SHA pinning, and a
few cross-file/policy items) are tracked in REMAINING.md.

+++ src/core/nexus/api.cpp
@@ -609,7 +609,7 @@ std::string Api::getApiKey()
   return api_key_;
 }
 
-std::optional<std::pair<std::string, int>> Api::extractDomainAndModId(const std::string& mod_url)
+std::optional<std::pair<std::string, long>> Api::extractDomainAndModId(const std::string& mod_url)
 {
   const std::regex regex(R"((?:https:\/\/)?www\.nexusmods\.com\/(.+)\/mods\/(\d+).*)");
   std::smatch match;
@@ -620,9 +620,11 @@ std::optional<std::pair<std::string, int>> Api::extractDomainAndModId(const std:
     // game-domain token so a crafted URL cannot redirect the call to another endpoint.
     if(!std::regex_match(domain, std::regex("[a-zA-Z0-9]+")))
       return {};
+    // Parse the mod id as a long so large Nexus ids are not truncated, and guard against
+    // std::stoll throwing on overflow/invalid input.
     try
     {
-      return { { domain, std::stoi(match[2]) } };
+      return { { domain, std::stoll(match[2]) } };
     }
     catch(const std::exception&)
     {
